动态规划

+++ 63_不同路径 II/63_Unique Paths ii/Unique Paths ii.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 int uniquePathsWithObstacles(vector<vector<int>>& obstacleGrid) {
 	if (obstacleGrid.empty() || obstacleGrid[0].empty() || obstacleGrid[0][0] == 1) return 0;
-	vector<vector<int> > dp(obstacleGrid.size(), vector<int>(obstacleGrid[0].size(), 0));
+	vector<vector<long> > dp(obstacleGrid.size(), vector<long>(obstacleGrid[0].size(), 0));
 	for (int i = 0; i < obstacleGrid.size(); ++i) {
 		for (int j = 0; j < obstacleGrid[i].size(); ++j) {
 			if (obstacleGrid[i][j] == 1) dp[i][j] = 0;
